python recieves and processes title of removed song
need to add checking to see if the correct song is being removed

+++ JukeBot/utils.py
@@ -12,6 +12,9 @@ def do_format(message):
     return endMsg
     
 def delete_file(dir):
+    if dir == 'bad_path' || dir == '':
+        print('No file exists - badpath/""')
+        return
     for x in range(30):
         try:
             os.unlink(dir)
